Controllers and routes refactored to users and thoughts.
